LimitSmall component rewritten to TS
