+++ src/components/LimitSmall/LimitSmall.mdx
@@ -1,5 +1,5 @@
 import { Playground, Props } from 'docz'
-import LimitSmall from '@components/LimitSmall'
+import { LimitSmall } from '@components/LimitSmall'
 
 # LimitSmall
