--- src/components/LimitSmall/LimitSmall.test.js
@@ -1,32 +1,0 @@
-import LimitSmall from '@components/LimitSmall'
-import { render, screen } from '@testing-library/react'
-import React from 'react'
-
-describe('<LimitSmall /> mount', () => {
-  const props = {
-    limit: 20000,
-    limitText: 'word.unique.visitors',
-    padding: 'none',
-    quantity: 5
-  }
-
-  it('is mounted', () => {
-    render(<LimitSmall {...props} />)
-  })
-})
-
-describe('<LimitSmall /> mount', () => {
-  const props = {
-    limit: -1,
-    limitText: 'word.unique.visitors',
-    padding: 'none',
-    quantity: 5
-  }
-
-  it('is mounted', () => {
-    render(<LimitSmall {...props} />)
-
-    const text = screen.getByText('/ ∞')
-    expect(text).toBeInDocument
-  })
-})
