+++ src/components/index.ts
@@ -50,7 +50,7 @@ export * from './Label'
 export * from './Layout'
 export { default as Legend } from './Legend'
 export * from './LimitQuantity'
-export { default as LimitSmall } from './LimitSmall'
+export * from './LimitSmall'
 export * from './List'
 export * from './List/Item'
 export { default as Loader } from './Loader'
